requirements.txt
Django==3.1
djangorestframework==3.11.1
image==1.5.32
numpy==1.16.2
scipy==1.5.2
scikit-learn==0.20.3
joblib==0.13.2